mgr/cephadm: moving service_discovery.py to services module
