+++ src/pybind/mgr/cephadm/http_server.py
@@ -4,7 +4,7 @@ import logging
 from typing import TYPE_CHECKING
 
 from cephadm.agent import AgentEndpoint
-from cephadm.service_discovery import ServiceDiscovery
+from cephadm.services.service_discovery import ServiceDiscovery
 from mgr_util import test_port_allocation, PortAlreadyInUse
 from orchestrator import OrchestratorError
