+++ src/pybind/mgr/cephadm/services/service_discovery.py
